Updated based on feedback from PR
Details:

* Updated some names
* Forwarded request body to the event of the invoke
* Added functional tests for echoing the event sent to lambda
* More refactoring around shared code
